requirements.txt
opencv-python
Flask
numpy
python-socketio
Flask-SocketIO
eventlet
waitress
cmake
dlib
face_recognition